extracted stepper component;
