+++ src/components/number-input/number-input.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import styles from './number-input.st.css';
 import inputStyles from '../../style/forms/input.st.css';
 import {SBComponent} from 'stylable-react-component';
-import {Stepper} from './stepper';
+import {Stepper} from '../stepper';
 import {KeyCodes} from '../../common/key-codes';
 
 export interface NumberInputProps {
